refactor: transfer Race3000Game class to app component
