+++ src/lib/types.ts
@@ -1,42 +1,3 @@
-export class Race3000Game {
-  public currentPlayer: Player;
-  public currentDiceFace: number;
-
-  constructor(
-    public players: Player[],
-    public dice: Dice
-  ) {
-    this.currentPlayer = players[0];
-    this.currentDiceFace = 0;
-  }
-
-  // Set next player's turn
-  nextPlayer(): void {
-    const currentIndex = this.players.indexOf(this.currentPlayer);
-    const nextIndex = (currentIndex + 1) % this.players.length;
-
-    this.currentPlayer = this.players[nextIndex];
-  }
-
-  // Remove a player from the game (typically as they've already finished the lap)
-  removePlayer(playerToRemove: Player): void {
-    this.players = this.players.filter((player) => player !== playerToRemove);
-  }
-
-  // Roll the dice, return a face
-  rollDice(): void {
-    const randomIndex = Math.floor(Math.random() * this.dice.length);
-    console.log("ROLLED: ", randomIndex);
-    this.currentDiceFace = randomIndex;
-  }
-
-  replacePieceOnCurrentFace(pieceIndex: number): void {
-    const movementPiece = new MovementPiece(this.currentPlayer);
-    this.dice[this.currentDiceFace][pieceIndex] = movementPiece;
-    return;
-  }
-}
-
 // Possible player names
 export const PLAYER_NAMES = ["red", "green", "blue", "white"];
